Remove controllers tests and move testing logic to domain models tests
